python/requirements.txt
hitherdither==0.1.6
numpy==1.19.4
Pillow==8.0.1
